Add payback chart components and implement scroll functionality
